backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
google-generativeai==0.3.2
elevenlabs==0.2.26
requests==2.31.0
python-jose[cryptography]==3.3.0
python-decouple==3.8
pydantic>=2.5.0
sqlalchemy==2.0.23
alembic==1.13.0
pillow>=10.4.0
aiofiles==23.2.0
